balloon: change intervals and thresholds to u32
kernel stores these values as u32 isntead of u64, so to prevent needless
casting change to use the same type in crosvm's balloon device.

BUG=b/285581266
TEST=cargo build && cargo build --no-default-features && cargo build --features registered_events
TEST=cargo build -p crosvm_control --features registered_events
TEST=cros_sdk emerge-volteer crosvm && cros_sdk emerge-volteer vm_host_tools

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4724907
Commit-Queue: maciek swiech <drmasquatch@google.com>

+++ common/balloon_control/src/lib.rs
@@ -27,9 +27,9 @@ pub enum BalloonTubeCommand {
     WorkingSet,
     // Send balloon ws config to guest.
     WorkingSetConfig {
-        bins: Vec<u64>,
-        refresh_threshold: u64,
-        report_threshold: u64,
+        bins: Vec<u32>,
+        refresh_threshold: u32,
+        report_threshold: u32,
     },
 }
