+++ devices/src/virtio/balloon.rs
@@ -598,9 +598,9 @@ async fn handle_events_queue(
 enum WSOp {
     WSReport,
     WSConfig {
-        bins: Vec<u64>,
-        refresh_threshold: u64,
-        report_threshold: u64,
+        bins: Vec<u32>,
+        refresh_threshold: u32,
+        report_threshold: u32,
     },
 }
